Delete LegacyUsersQuestionnaireOnboardingComponent
