--- iosHyperskillApp/iosHyperskillApp/Sources/Modules/UsersQuestionnaireOnboarding/UsersQuestionnaireOnboardingOutputProtocol.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol UsersQuestionnaireOnboardingOutputProtocol: AnyObject {
-    func handleUsersQuestionnaireOnboardingCompleted()
-}
