+++ shared/src/commonMain/kotlin/org/hyperskill/app/welcome_onboarding/finish/injection/WelcomeOnboardingFinishComponentImpl.kt
@@ -3,7 +3,7 @@ package org.hyperskill.app.welcome_onboarding.finish.injection
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.welcome_onboarding.finish.view.WelcomeOnboardingFinishViewStateMapper
 
-class WelcomeOnboardingFinishComponentImpl(
+internal class WelcomeOnboardingFinishComponentImpl(
     private val appGraph: AppGraph
 ) : WelcomeOnboardingFinishComponent {
     override val welcomeOnboardingFinishViewStateMapper: WelcomeOnboardingFinishViewStateMapper
