+++ shared/src/commonMain/kotlin/org/hyperskill/app/welcome_onboarding/questionnaire/injection/WelcomeQuestionnaireComponentImpl.kt
@@ -3,12 +3,12 @@ package org.hyperskill.app.welcome_onboarding.questionnaire.injection
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.welcome_onboarding.questionnaire.view.WelcomeQuestionnaireViewStateMapper
 
-class WelcomeQuestionnaireComponentImpl(
+internal class WelcomeQuestionnaireComponentImpl(
     private val appGraph: AppGraph
 ) : WelcomeQuestionnaireComponent {
     override val welcomeQuestionnaireViewStateMapper: WelcomeQuestionnaireViewStateMapper
         get() = WelcomeQuestionnaireViewStateMapper(
-            resourceProvider = appGraph.commonComponent.resourceProvider,
-            platformType = appGraph.commonComponent.platform.platformType
+            platform = appGraph.commonComponent.platform,
+            resourceProvider = appGraph.commonComponent.resourceProvider
         )
 }
